updated repo structure, pom.xml test
testing Maven build
