+++ src/main/java/edu/wit/comp2000/group23/application3/IOccupant.java
